Added throwing exception if not found in CQRS DeleteGoodToCategoryCommandHandler

+++ Atlas.Application/CQRS/CategoryToGoods/Commands/DeleteCategoryToGood/DeleteCategoryToGoodCommandHandler.cs
@@ -1,10 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Atlas.Application.Common.Exceptions;
 using Atlas.Application.Interfaces;
+using Atlas.Domain;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
@@ -21,6 +20,11 @@ namespace Atlas.Application.CQRS.CategoryToGoods.Commands.DeleteCategoryToGood
         {
             var categoryToGood = await _dbContext.CategoryToGoods
                 .FirstOrDefaultAsync(ctg => ctg.Id == request.Id, cancellationToken);
+
+            if (categoryToGood == null)
+            {
+                throw new NotFoundException(nameof(CategoryToGood), request.Id);
+            }
 
             _dbContext.CategoryToGoods.Remove(categoryToGood);
             await _dbContext.SaveChangesAsync(cancellationToken);
